Scenarios fixes: * Added additional methods * Changed waiter method
