+++ src/test/resources/features/SearchFuntionality.feature
@@ -1,9 +1,9 @@
+@aciuvaga
 Feature: Search functionality
 
   Background:
     Given user is on Home page
 
-  @aciuvaga
   Scenario Outline: Search a product by keyword
     When Search store is populated with '<search_Item>'
     And 'SEARCH' button is pressed
